updated store, created allMugs and oneMug reducers, thunks
