+++ client/store/index.js
@@ -2,9 +2,15 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
+import userReducer from './user'
+import oneMugReducer from './oneMug'
+import allMugsReducer from './allMugs'
 
-const reducer = combineReducers({user})
+const reducer = combineReducers({
+  user: userReducer,
+  oneMug: oneMugReducer,
+  allMugs: allMugsReducer
+})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
